implemented filter function to top metrics
