+++ Components/Pages/Goals.js
@@ -19,64 +19,8 @@ import { AsyncStorage } from "react-native";
 import { NavigationEvents } from "react-navigation";
 import AnimatedHeader from "../Molecules/AnimatedHeader";
 import _ from "lodash";
-import { determineOverDue } from "../Atoms/BarChart.functions";
+import { GoalsSort, GoalsFilterState, GoalsFilterCadence } from "../Atoms/BarChart.functions";
 
-const GoalsSort = ({ goals }) => {
-  const [sortOrder, updateSortOrder] = useState("none");
-  const goals_copy = [...goals];
-
-  if (sortOrder == "none") {
-    var sorted_goals = goals_copy;
-    return { sorted_goals, updateSortOrder, sortOrder };
-  } else if (sortOrder == "asc") {
-    var sorted_goals = _.sortBy(goals_copy, function(D) {
-      return D["timeStamps"].length;
-    });
-    return { sorted_goals, updateSortOrder, sortOrder };
-  } else {
-    var sorted_goals = _.sortBy(goals_copy, function(D) {
-      return D["timeStamps"].length * -1;
-    });
-    return { sorted_goals, updateSortOrder, sortOrder };
-  }
-};
-
-const GoalsFilterState = ({ goals, state }) => {
-  if (state == "all") {
-    return goals;
-  } else if (state == "incomplete") {
-    var filtered_goals = goals.filter(function(D) {
-      return determineOverDue({ ...D, goals });
-    });
-    return filtered_goals;
-  } else if (state == "complete") {
-    var filtered_goals = goals.filter(function(D) {
-      return !determineOverDue({ ...D, goals });
-    });
-    return filtered_goals;
-  }
-};
-
-const GoalsFilterCadence = ({ goals, cadence }) => {
-  if (cadence == "all") {
-    return goals;
-  } else if (cadence == "daily") {
-    var filtered_goals = goals.filter(function(D) {
-      return String(D["cadence"]).toLowerCase() == "daily";
-    });
-    return filtered_goals;
-  } else if (cadence == "weekly") {
-    var filtered_goals = goals.filter(function(D) {
-      return String(D["cadence"]).toLowerCase() == "weekly";
-    });
-    return filtered_goals;
-  } else if (cadence == "monthly") {
-    var filtered_goals = goals.filter(function(D) {
-      return String(D["cadence"]).toLowerCase() == "monthly";
-    });
-    return filtered_goals;
-  }
-};
 const GoalsFilter = ({ goals }) => {
   const [filter, updateFilter] = useState({ state: "all", cadence: "all" });
